requirements_api.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
python-magic-bin==0.4.14
pillow-heif==0.18.0
gunicorn==21.2.0